fix(textfield): вынес в отдельный компонент
Jira issue: VEGA-834
